Move functions to new file
