+++ registry/src/lib.rs
@@ -4,6 +4,7 @@ use sqlx::{ConnectOptions, Pool, Postgres};
 pub mod config;
 pub mod download;
 pub mod errors;
+pub mod extensions;
 pub mod routes;
 pub mod token;
 pub mod uploader;
